requirements.txt
pandas
networkx
haversine
FuzzyWuzzy
staticmap
python-telegram-bot
requests
python-Levenshtein
